+++ quiltcore/__init__.py
@@ -25,6 +25,7 @@ manages the 'quiltcore.yaml' configuration file.
 """
 
 from .builder import Builder  # noqa: F401
+from .builder2 import FolderBuilder  # noqa: F401
 from .changes import Changes  # noqa: F401
 from .delta import Delta  # noqa: F401
 from .domain import Domain  # noqa: F401