+++ quiltcore/udg/folder.py
@@ -8,13 +8,14 @@ from .node import Node
 class Folder(Child):
     KEY_DIR = "quilt3/dirs/"
     KEY_GLOB = "glob"
+    DEFAULT_GLOB = "*"
 
     def __init__(self, name: str, parent: Node, **kwargs):
         super().__init__(name, parent, **kwargs)
 
     def _setup(self):
         super()._setup()
-        self.glob = self.param(self.KEY_GLOB, "*")
+        self.glob = self.param(self.KEY_GLOB, self.DEFAULT_GLOB)
 
     def _setup_dir(self, path: Path, key: str) -> Path:
         """Form dir and create if it does not exist."""