+++ tests/test_pull.py
@@ -7,12 +7,17 @@ from quiltcore import UDI, Domain, Scheme, quilt
 from .conftest import LOCAL_UDI, LOCAL_URI, LOCAL_VOL, TEST_HASH, TEST_PKG, not_win
 
 
-@pytest.fixture
-def domain():
+def make_domain():
     with TemporaryDirectory() as tmpdirname:
         f = quilt["file"]
         dom = f[tmpdirname]
         dom.is_mutable = True
+        yield dom
+
+
+@pytest.fixture
+def domain():
+    for dom in make_domain():
         yield dom
 
 
@@ -30,12 +35,28 @@ def test_pull_udi_scheme():
     assert LOCAL_VOL in dom.cf.AsStr(dom.store)
 
 
-def test_pull_udi(domain: Domain, remote_udi: UDI):
+def test_pull_remote_udi(remote_udi: UDI):
     assert remote_udi
     assert remote_udi.uri == LOCAL_UDI
     assert remote_udi.package == TEST_PKG
     if not_win():
         assert remote_udi.registry == LOCAL_URI
+
+
+def test_pull_uri(domain: Domain):
+    uri = domain.get_uri()
+    assert uri.startswith("file:///")
+    domain2 = Domain.FromURI(uri)
+    assert domain2.store == domain.store
+
+
+def test_pull_udi(domain: Domain):
+    udi = domain.get_udi_string()
+    assert udi.startswith("quilt+file:///")
+    udip = domain.get_udi_string(TEST_PKG)
+    assert udip.endswith(f"#package={TEST_PKG}")
+    udipp = domain.get_udi_string(TEST_PKG, "README.md")
+    assert udipp.endswith(f"#package={TEST_PKG}&path=README.md")
 
 
 def test_pull_raise(domain: Domain, remote_udi: UDI):
@@ -54,11 +75,26 @@ def test_pull_data_yaml(domain: Domain, remote_udi: UDI):
 
     dest = str(path)
     assert TEST_PKG in dest
-    assert dy.get_uri(dest) == LOCAL_UDI
-    assert dy.get_folder(LOCAL_UDI) == dest
+    assert dy.folder2uri(dest) == LOCAL_UDI
+    assert dy.uri2folder(LOCAL_UDI) == dest
 
     status = dy.get_list(dest, LOCAL_UDI, "pull")
     assert isinstance(status, dict)
     assert "timestamp" in status
     assert "user" in status
     assert status["hash"] == TEST_HASH
+
+
+# @pytest.mark.skip(reason="Fix user_meta encoding")
+def test_pull_push():
+    for local in make_domain():
+        local_path = local.package_path(TEST_PKG)
+        assert local_path.exists()
+        readme = local_path / "README.md"
+        readme.write_text("Hello World")
+        local.commit(local_path, package=TEST_PKG, msg=f"test_pull_push {Domain.Now()}")
+        # local.push(local_path)
+    for remote in make_domain():
+        assert isinstance(remote, Domain)
+        remote_udi = remote.get_udi(TEST_PKG)
+        assert remote_udi